+++ nvim/.config/nvim/colors/pepe.lua
@@ -1,0 +1,5 @@
+package.loaded["pepe"] = nil
+package.loaded["pepe.base"] = nil
+package.loaded["pepe.plugins.highlights"] = nil
+package.loaded["pepe.languages"] = nil
+require("pepe").load()