+++ nvim/.config/nvim/lua/netzego/gitsigns.lua
@@ -1,1 +1,2 @@
+-- https://github.com/lewis6991/gitsigns.nvim
 require("gitsigns").setup()