+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 	- Add gitsigns
   - Add fugitive
   - Add lualine
+* 2022-02-20
+  - Move lua/config to lua/netzego
+  -